@@ -36,108 +36,107 @@
     ]
   },
   "dependencies": {
-    "@babel/runtime-corejs3": "^7.12.5",
-    "@emotion/react": "^11.1.1",
-    "@material-ui/core": "^4.11.1",
-    "@material-ui/icons": "^4.9.1",
-    "@react-leaflet/core": "1.0.2",
-    "@svgr/webpack": "^5.5.0",
-    "@types/lodash.throttle": "^4.1.6",
-    "@types/node": "^14.14.20",
-    "@types/react-redux": "^7.1.16",
-    "@types/react-router": "^5.1.11",
-    "amazon-cognito-identity-js": "^4.5.6",
+    "@babel/runtime-corejs3": "^7.18.3",
+    "@emotion/react": "^11.9.0",
+    "@material-ui/core": "^4.12.4",
+    "@material-ui/icons": "^4.11.3",
+    "@react-leaflet/core": "2.0.0",
+    "@svgr/webpack": "^6.2.1",
+    "@types/lodash.throttle": "^4.1.7",
+    "@types/node": "^17.0.35",
+    "@types/react-redux": "^7.1.24",
+    "@types/react-router": "^5.1.18",
+    "amazon-cognito-identity-js": "^5.2.9",
     "amplify": "0.0.11",
     "aws-sdk": "https://github.com/lucastrcts/aws-sdk-js",
-    "axios": "^0.21.4",
-    "connected-react-router": "^6.8.0",
+    "axios": "^0.27.2",
+    "connected-react-router": "^6.9.2",
     "crypto-browserify": "^3.12.0",
-    "esri-leaflet": "2.5.1",
-    "esri-leaflet-renderers": "2.0.6",
-    "i18next": "19.8.4",
-    "i18next-http-backend": "^1.3.2",
-    "immutable": "^4.0.0-rc.12",
-    "leaflet": "1.7.1",
+    "esri-leaflet": "3.0.8",
+    "i18next": "21.8.4",
+    "i18next-http-backend": "^1.4.1",
+    "immutable": "^4.1.0",
+    "leaflet": "1.8.0",
     "lodash.throttle": "^4.1.1",
     "proj4leaflet": "1.0.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-helmet": "^6.1.0",
-    "react-i18next": "11.8.0",
-    "react-leaflet": "3.0.2",
-    "react-redux": "^7.2.2",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
+    "react-i18next": "11.16.9",
+    "react-leaflet": "4.0.0",
+    "react-redux": "^8.0.2",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
     "react-share": "^4.4.0",
-    "react-spinners": "^0.10.4",
-    "reactstrap": "^8.8.1",
-    "redux": "^4.0.5",
-    "screenfull": "~5.0.2"
+    "react-spinners": "^0.12.0",
+    "reactstrap": "^9.0.3",
+    "redux": "^4.2.0",
+    "screenfull": "~6.0.1"
   },
   "devDependencies": {
-    "@babel/cli": "^7.12.8",
-    "@babel/core": "^7.12.9",
-    "@babel/plugin-proposal-class-properties": "^7.12.1",
-    "@babel/plugin-proposal-decorators": "^7.12.1",
-    "@babel/plugin-proposal-do-expressions": "^7.12.1",
-    "@babel/plugin-proposal-export-default-from": "^7.12.1",
-    "@babel/plugin-proposal-export-namespace-from": "^7.12.1",
-    "@babel/plugin-proposal-function-bind": "^7.12.1",
-    "@babel/plugin-proposal-function-sent": "^7.12.1",
-    "@babel/plugin-proposal-json-strings": "^7.12.1",
-    "@babel/plugin-proposal-logical-assignment-operators": "^7.12.1",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.12.1",
-    "@babel/plugin-proposal-numeric-separator": "^7.12.7",
-    "@babel/plugin-proposal-optional-chaining": "^7.12.7",
-    "@babel/plugin-proposal-pipeline-operator": "^7.12.1",
-    "@babel/plugin-proposal-private-methods": "^7.12.1",
-    "@babel/plugin-proposal-throw-expressions": "^7.12.1",
+    "@babel/cli": "^7.17.10",
+    "@babel/core": "^7.18.2",
+    "@babel/plugin-proposal-class-properties": "^7.17.12",
+    "@babel/plugin-proposal-decorators": "^7.18.2",
+    "@babel/plugin-proposal-do-expressions": "^7.16.7",
+    "@babel/plugin-proposal-export-default-from": "^7.17.12",
+    "@babel/plugin-proposal-export-namespace-from": "^7.17.12",
+    "@babel/plugin-proposal-function-bind": "^7.16.7",
+    "@babel/plugin-proposal-function-sent": "^7.18.2",
+    "@babel/plugin-proposal-json-strings": "^7.17.12",
+    "@babel/plugin-proposal-logical-assignment-operators": "^7.17.12",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.17.12",
+    "@babel/plugin-proposal-numeric-separator": "^7.16.7",
+    "@babel/plugin-proposal-optional-chaining": "^7.17.12",
+    "@babel/plugin-proposal-pipeline-operator": "^7.18.2",
+    "@babel/plugin-proposal-private-methods": "^7.17.12",
+    "@babel/plugin-proposal-throw-expressions": "^7.16.7",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
     "@babel/plugin-syntax-import-meta": "^7.10.4",
-    "@babel/plugin-transform-runtime": "^7.12.1",
-    "@babel/preset-env": "^7.12.7",
-    "@babel/preset-react": "^7.12.7",
-    "@babel/preset-typescript": "^7.12.7",
-    "@types/esri-leaflet": "^2.1.6",
-    "@types/flux": "^3.1.9",
-    "@types/jest": "^26.0.16",
-    "@types/leaflet": "1.5.19",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-leaflet": "2.5.2",
-    "@types/react-router-dom": "^5.1.6",
-    "@typescript-eslint/eslint-plugin": "^4.9.0",
-    "@typescript-eslint/parser": "^4.9.0",
+    "@babel/plugin-transform-runtime": "^7.18.2",
+    "@babel/preset-env": "^7.18.2",
+    "@babel/preset-react": "^7.17.12",
+    "@babel/preset-typescript": "^7.17.12",
+    "@types/esri-leaflet": "^2.1.9",
+    "@types/flux": "^3.1.11",
+    "@types/jest": "^27.5.1",
+    "@types/leaflet": "1.7.10",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.5",
+    "@types/react-leaflet": "2.8.2",
+    "@types/react-router-dom": "^5.3.3",
+    "@typescript-eslint/eslint-plugin": "^5.26.0",
+    "@typescript-eslint/parser": "^5.26.0",
     "babel-eslint": "^10.1.0",
-    "babel-jest": "^26.6.3",
-    "babel-loader": "^8.2.2",
-    "babel-plugin-import": "1.13.3",
-    "copy-webpack-plugin": "^7.0.0",
-    "css-loader": "5.0.1",
-    "eslint": "^7.14.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-config-prettier": "^6.15.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jest": "^24.1.3",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.21.5",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "babel-jest": "^28.1.0",
+    "babel-loader": "^8.2.5",
+    "babel-plugin-import": "1.13.5",
+    "copy-webpack-plugin": "^11.0.0",
+    "css-loader": "6.7.1",
+    "eslint": "^8.16.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest": "^26.2.2",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.30.0",
+    "eslint-plugin-react-hooks": "^4.5.0",
     "file-loader": "6.2.0",
-    "gh-pages": "2.1.1",
-    "html-webpack-plugin": "^5.0.0-alpha.14",
-    "jest": "^26.6.3",
-    "prettier": "^2.2.1",
-    "react-test-renderer": "^17.0.1",
-    "sass": "^1.23.7",
-    "sass-loader": "^8.0.0",
-    "style-loader": "2.0.0",
-    "typescript": "^4.1.2",
-    "webpack": "^5.9.0",
-    "webpack-bundle-analyzer": "^4.2.0",
+    "gh-pages": "4.0.0",
+    "html-webpack-plugin": "^5.5.0",
+    "jest": "^28.1.0",
+    "prettier": "^2.6.2",
+    "react-test-renderer": "^18.1.0",
+    "sass": "^1.52.1",
+    "sass-loader": "^13.0.0",
+    "style-loader": "3.3.1",
+    "typescript": "^4.7.2",
+    "webpack": "^5.72.1",
+    "webpack-bundle-analyzer": "^4.5.0",
     "webpack-bundle-size-analyzer": "^3.1.0",
-    "webpack-cli": "^4.4.0",
-    "webpack-dev-server": "^3.11.2",
-    "webpack-merge": "^5.4.0"
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.9.0",
+    "webpack-merge": "^5.8.0"
   }
 }
